requirements.txt
torch
torchvision
opencv-python
matplotlib
numpy
pillow
opendatasets